Moving range and RangeBuilder into their own folder
