+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "lint": "./node_modules/.bin/eslint src",
     "test": "npm run lint && nyc npm run run-tests",
-    "run-tests": "babel-tape-runner **/*.test.js | faucet",
+    "run-tests": "babel-tape-runner **/__test__/*.test.js | faucet",
     "build": "rm -rf ./lib && babel src --out-dir lib --ignore test.js",
     "prepare": "npm run test && npm run build"
   },
@@ -37,6 +37,7 @@
   },
   "homepage": "https://github.com/hillc5/range-iterator#readme",
   "devDependencies": {
+    "@sinonjs/referee-sinon": "^5.0.0",
     "babel-cli": "^6.26.0",
     "babel-eslint": "^8.2.6",
     "babel-plugin-transform-object-rest-spread": "^6.26.0",
@@ -44,14 +45,14 @@
     "babel-preset-env": "^1.7.0",
     "babel-tape-runner": "^2.0.1",
     "eslint": "^4.19.1",
-    "eslint-config-airbnb-base": "^13.0.0",
-    "eslint-plugin-import": "^2.13.0",
+    "eslint-config-airbnb-base": "^13.1.0",
+    "eslint-plugin-import": "^2.17.2",
     "faucet": "0.0.1",
-    "nyc": "^12.0.2",
-    "sinon": "^6.1.3",
-    "tape": "^4.9.1"
+    "nyc": "^14.1.1",
+    "sinon": "^6.3.5",
+    "tape": "^4.10.1"
   },
   "dependencies": {
-    "regenerator-runtime": "^0.12.0"
+    "regenerator-runtime": "^0.12.1"
   }
 }
